[Fuzzy Search] Rename `find` to `exactFind` and `fuzzyFind` to `find`
