+++ src/main/java/organice/logic/parser/AddressBookParser.java
@@ -6,20 +6,8 @@ import static organice.commons.core.Messages.MESSAGE_UNKNOWN_COMMAND;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
-import organice.logic.commands.AddCommand;
-import organice.logic.commands.ClearCommand;
-import organice.logic.commands.Command;
-import organice.logic.commands.DoneCommand;
-import organice.logic.commands.EditCommand;
-import organice.logic.commands.ExitCommand;
-import organice.logic.commands.FindCommand;
-import organice.logic.commands.FuzzyFindCommand;
-import organice.logic.commands.HelpCommand;
-import organice.logic.commands.ListCommand;
-import organice.logic.commands.MatchCommand;
-import organice.logic.commands.ProcessingCommand;
-import organice.logic.commands.ProcessingMarkDoneCommand;
-import organice.logic.commands.SortCommand;
+import organice.logic.commands.*;
+import organice.logic.commands.ExactFindCommand;
 import organice.logic.parser.exceptions.ParseException;
 
 /**
@@ -58,11 +46,11 @@ public class AddressBookParser {
         case ClearCommand.COMMAND_WORD:
             return new ClearCommand();
 
-        case FuzzyFindCommand.COMMAND_WORD:
-            return new FuzzyFindCommandParser().parse(arguments);
-
         case FindCommand.COMMAND_WORD:
             return new FindCommandParser().parse(arguments);
+
+        case ExactFindCommand.COMMAND_WORD:
+            return new ExactFindCommandParser().parse(arguments);
 
         case ListCommand.COMMAND_WORD:
             return new ListCommandParser().parse(arguments);
