+++ src/main/java/organice/logic/parser/DoneCommandParser.java
@@ -9,7 +9,7 @@ import organice.logic.commands.DoneCommand;
 import organice.logic.parser.exceptions.ParseException;
 
 /**
- * Parses input arguments and creates a new FindCommand object
+ * Parses input arguments and creates a new ExactFindCommand object
  */
 public class DoneCommandParser implements Parser<DoneCommand> {
