+++ src/main/java/organice/logic/parser/FindCommandParser.java
@@ -15,16 +15,15 @@ import static organice.logic.parser.CliSyntax.PREFIX_TYPE;
 
 import organice.logic.commands.FindCommand;
 import organice.logic.parser.exceptions.ParseException;
-import organice.model.person.PersonContainsPrefixesPredicate;
 
 /**
- * Parses input arguments and creates a new FindCommand object
+ * Parses input arguments and creates a new ExactFindCommand object
  */
 public class FindCommandParser implements Parser<FindCommand> {
 
     /**
-     * Parses the given {@code String} of arguments in the context of the FindCommand
-     * and returns a FindCommand object for execution.
+     * Parses the given {@code String} of arguments in the context of the ExactFindCommand
+     * and returns a ExactFindCommand object for execution.
      * @throws ParseException if the user input does not conform the expected format
      */
     public FindCommand parse(String args) throws ParseException {
@@ -34,12 +33,12 @@ public class FindCommandParser implements Parser<FindCommand> {
                 PREFIX_TYPE, PREFIX_AGE, PREFIX_PRIORITY, PREFIX_BLOOD_TYPE, PREFIX_DOCTOR_IN_CHARGE,
                 PREFIX_TISSUE_TYPE, PREFIX_ORGAN_EXPIRY_DATE, PREFIX_ORGAN);
 
-        // Preamble length > 4 indicates initial command is not `FindCommand.COMMAND_WORD PREFIX/KEYWORD`
-        if (trimmedArgs.isEmpty() || argMultimap.getPreamble().length() > FindCommand.COMMAND_WORD.length()) {
+        // Preamble length > 7 indicates initial command is not `fuzfind PREFIX/KEYWORD`
+        if (trimmedArgs.isEmpty() || argMultimap.getPreamble().length() > 7) {
             throw new ParseException(
                     String.format(MESSAGE_INVALID_COMMAND_FORMAT, FindCommand.MESSAGE_USAGE));
         }
-        return new FindCommand(new PersonContainsPrefixesPredicate(argMultimap));
+        return new FindCommand(argMultimap);
     }
 
 }
