+++ src/main/java/organice/logic/parser/ProcessingCommandParser.java
@@ -9,7 +9,7 @@ import organice.logic.commands.ProcessingCommand;
 import organice.logic.parser.exceptions.ParseException;
 
 /**
- * Parses input arguments and creates a new FindCommand object
+ * Parses input arguments and creates a new ExactFindCommand object
  */
 public class ProcessingCommandParser implements Parser<ProcessingCommand> {
