+++ src/test/java/organice/model/person/PersonContainsPrefixesPredicateTest.java
@@ -16,7 +16,7 @@ import static organice.logic.parser.CliSyntax.PREFIX_TYPE;
 
 import org.junit.jupiter.api.Test;
 
-import organice.logic.commands.FindCommand;
+import organice.logic.commands.ExactFindCommand;
 import organice.logic.parser.ArgumentMultimap;
 import organice.logic.parser.ArgumentTokenizer;
 import organice.testutil.PersonBuilder;
@@ -27,7 +27,7 @@ public class PersonContainsPrefixesPredicateTest {
     public void equals() {
         //TODO: Replace ArgumentTokenizer with stub
         ArgumentMultimap firstPredicateKeywordList =
-                ArgumentTokenizer.tokenize(FindCommand.COMMAND_WORD + " n/Alice", PREFIX_NAME);
+                ArgumentTokenizer.tokenize(ExactFindCommand.COMMAND_WORD + " n/Alice", PREFIX_NAME);
         ArgumentMultimap secondPredicateKeywordList = ArgumentTokenizer
                 .tokenize("n/Alice Benson", PREFIX_NAME);
 
@@ -58,22 +58,22 @@ public class PersonContainsPrefixesPredicateTest {
     public void test_personContainsPrefixes_returnsTrue() {
         // One keyword
         PersonContainsPrefixesPredicate predicate = new PersonContainsPrefixesPredicate(
-                        ArgumentTokenizer.tokenize(FindCommand.COMMAND_WORD + " n/Alice", PREFIX_NAME));
+                        ArgumentTokenizer.tokenize(ExactFindCommand.COMMAND_WORD + " n/Alice", PREFIX_NAME));
         assertTrue(predicate.test(new PersonBuilder().withName("Alice Bob").build()));
 
         // Multiple keywords
         predicate = new PersonContainsPrefixesPredicate(
-                ArgumentTokenizer.tokenize(FindCommand.COMMAND_WORD + " n/Alice Bob", PREFIX_NAME));
+                ArgumentTokenizer.tokenize(ExactFindCommand.COMMAND_WORD + " n/Alice Bob", PREFIX_NAME));
         assertTrue(predicate.test(new PersonBuilder().withName("Alice Bob").build()));
 
         // Only one matching keyword
         predicate = new PersonContainsPrefixesPredicate(
-                ArgumentTokenizer.tokenize(FindCommand.COMMAND_WORD + " n/Bob Carol", PREFIX_NAME));
+                ArgumentTokenizer.tokenize(ExactFindCommand.COMMAND_WORD + " n/Bob Carol", PREFIX_NAME));
         assertFalse(predicate.test(new PersonBuilder().withName("Alice Carol").build()));
 
         // Mixed-case keywords
         predicate = new PersonContainsPrefixesPredicate(
-                ArgumentTokenizer.tokenize(FindCommand.COMMAND_WORD + " n/aLIce bOB", PREFIX_NAME));
+                ArgumentTokenizer.tokenize(ExactFindCommand.COMMAND_WORD + " n/aLIce bOB", PREFIX_NAME));
         assertTrue(predicate.test(new PersonBuilder().withName("Alice Bob").build()));
     }
 
@@ -81,19 +81,19 @@ public class PersonContainsPrefixesPredicateTest {
     public void test_nameDoesNotContainKeywords_returnsFalse() {
         // Zero keywords should return false
         PersonContainsPrefixesPredicate predicate = new PersonContainsPrefixesPredicate(ArgumentTokenizer
-                .tokenize(FindCommand.COMMAND_WORD, PREFIX_NAME, PREFIX_NRIC, PREFIX_PHONE, PREFIX_TYPE,
+                .tokenize(ExactFindCommand.COMMAND_WORD, PREFIX_NAME, PREFIX_NRIC, PREFIX_PHONE, PREFIX_TYPE,
                         PREFIX_AGE, PREFIX_PRIORITY, PREFIX_BLOOD_TYPE, PREFIX_DOCTOR_IN_CHARGE, PREFIX_TISSUE_TYPE,
                         PREFIX_ORGAN_EXPIRY_DATE, PREFIX_ORGAN));
         assertFalse(predicate.test(new PersonBuilder().withName("Alice").build()));
 
         // Non-matching keyword
         predicate = new PersonContainsPrefixesPredicate(
-                ArgumentTokenizer.tokenize(FindCommand.COMMAND_WORD + " n/Carol", PREFIX_NAME));
+                ArgumentTokenizer.tokenize(ExactFindCommand.COMMAND_WORD + " n/Carol", PREFIX_NAME));
         assertFalse(predicate.test(new PersonBuilder().withName("Alice Bob").build()));
 
         // Keywords match type, phone, nric and address, but does not match name
         predicate = new PersonContainsPrefixesPredicate(ArgumentTokenizer
-                .tokenize(FindCommand.COMMAND_WORD + " n/Benson t/doctor ic/S1111111A p/12345",
+                .tokenize(ExactFindCommand.COMMAND_WORD + " n/Benson t/doctor ic/S1111111A p/12345",
                         PREFIX_NAME, PREFIX_TYPE, PREFIX_NRIC, PREFIX_PHONE));
         assertFalse(predicate.test(new PersonBuilder().withType("doctor").withNric("S1111111A").withName("Alice")
                 .withPhone("12345").build()));
